src/server/requirements.txt
sanic[ext]==23.12.1
tortoise-orm[accel]==0.20.0
aiomysql==0.2.0
asyncpg==0.29.0
sanic-openapi==21.12.0